module linux check for pwritev2, preadv2 linux version

+++ uclang/mods/linux_uclm/header_files/ucl_linux.h
@@ -41,7 +41,7 @@ struct mmap_s
 {
   location_s *fd_location;
   void *mem_ptr;
-  size_t length;
+  long long int length;
 
   inline void init();
   void clear(interpreter_thread_s &it);
